update orders by user modal, make into component user modal header, add copy user id button
