+++ src/components/AddCouponModal/AddCouponModal.css
@@ -27,7 +27,6 @@
   padding: 10px;
   border: 1px solid #ccc;
   font-size: 1rem;
-  border-radius: 4px;
 }
 
 .coupon-form-group input:focus {
