+++ src/components/ViewUserOrdersModal/ViewUserOrdersModal.css
@@ -4,7 +4,7 @@
   /* Modal content styles handled by Modal component */
 }
 
-.view-user-orders-header {
+/* .view-user-orders-header {
   margin-bottom: 24px;
   padding-bottom: 16px;
   border-bottom: 1px solid #e9ecef;
@@ -26,13 +26,13 @@
 .view-user-orders-last-login {
   color: #6c757d;
   font-size: 14px;
-}
+} */
 
 .view-user-orders-content {
   /* Content styles */
 }
 
-.view-user-orders-title {
+/* .view-user-orders-title {
   font-size: 16px;
   font-weight: 600;
   margin-bottom: 16px;
@@ -94,10 +94,10 @@
 .order-actions-button:hover {
   background-color: #f8f9fa;
   color: #333;
-}
+} */
 
 /* Responsive adjustments */
-@media (max-width: 768px) {
+/* @media (max-width: 768px) {
   .order-item {
     flex-direction: column;
     align-items: flex-start;
@@ -106,4 +106,4 @@
   .item-id {
     margin-top: 2px;
   }
-}
+} */
