+++ src/components/index.js
@@ -14,7 +14,7 @@ export { default as Modal } from "./Modal/Modal";
 export { default as UserContextMenu } from "./UserContextMenu/UserContextMenu";
 export { default as UserProfileModal } from "./UserProfileModal/UserProfileModal";
 export { default as ViewUserDetailsModal } from "./ViewUserDetailsModal/ViewUserDetailsModal";
-export { default as ViewUserOrdersModal } from "./ViewUserOrdersModal";
+export { default as ViewUserOrdersModal } from "./ViewUserOrdersModal/ViewUserOrdersModal";
 
 // Other Components
 export { default as ProtectedRoute } from "./ProtectedRoute";
